refactor(team-settings): create users tab in team settings
